+++ src/Security/EventSubscriber/SecuritySubscriber.php
@@ -39,7 +39,7 @@ class SecuritySubscriber implements EventSubscriberInterface
     /**
      * @inheritDoc
      */
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return [
             SecurityEvents::SWITCH_USER => 'onSwitchUser',